Rename files and add unit tests
